refactor: judgement를 dealer 내부로 통합
